+++ tests/testthat/test-createCluster.R
@@ -73,7 +73,21 @@ sapply(studies, function(study) {
         add_prefix = FALSE
       )
     }
-    expect_error(antaresRead::readClusterDesc())
+    
+    # list .ini files
+    path_thermal <- file.path(opts$inputPath, "thermal", "clusters")
+    all_ini_files <- list.files(path_thermal, full.names = TRUE, recursive = TRUE)
+    
+    # extract number or raw
+    suppressWarnings(
+      ini_nb_raw <- sapply(
+        lapply(all_ini_files, 
+               data.table::fread), 
+        function(x) dim(x)[2])
+    )
+    
+    # test all ini files are empty
+    expect_equal(sum(ini_nb_raw), 0)
     
   })