+++ NEWS.md
@@ -1,4 +1,7 @@
 # antaresEditObject 0.6.2
+* Fix test to remove rev dep to `antaresRead` (see `cran-comments.md` for details)
+
+# antaresEditObject 0.6.1
 
 NEW FEATURES :